아이템 unique키 작업, 비주얼 스튜디오 2022적용, db수정, 룸 락 개선

+++ GameServer/GameInit.cpp
@@ -55,12 +55,12 @@ void GameInit::SetMap(boost::json::value& mapJson, boost::json::value& mapUnitJs
 	for (int i = 0; i < mapJson.get_array().size(); i++)
 	{
 		auto& map = mapJson.get_array()[i];
-		int32 type = GameUtils::JsonParser::Parser("type", map).get_int64();
-		int32 code = GameUtils::JsonParser::Parser("mapCode", map).get_int64();
-		int32 x = GameUtils::JsonParser::Parser("x", map).get_int64();
-		int32 y = GameUtils::JsonParser::Parser("y", map).get_int64();
-		int32 centerX = GameUtils::JsonParser::Parser("centerX", map).get_int64();
-		int32 centerY = GameUtils::JsonParser::Parser("centerY", map).get_int64();
+		int32 type = static_cast<int32>(GameUtils::JsonParser::Parser("type", map).get_int64());
+		int32 code = static_cast<int32>(GameUtils::JsonParser::Parser("mapCode", map).get_int64());
+		int32 x = static_cast<int32>(GameUtils::JsonParser::Parser("x", map).get_int64());
+		int32 y = static_cast<int32>(GameUtils::JsonParser::Parser("y", map).get_int64());
+		int32 centerX = static_cast<int32>(GameUtils::JsonParser::Parser("centerX", map).get_int64());
+		int32 centerY = static_cast<int32>(GameUtils::JsonParser::Parser("centerY", map).get_int64());
 
 		GameRoomRef room = GRoomManger->CreateRoom(code);
 		CrashFunc(room != nullptr);
@@ -75,10 +75,10 @@ void GameInit::SetMap(boost::json::value& mapJson, boost::json::value& mapUnitJs
 		{
 			boost::json::value monsterMapJson = GameUtils::JsonParser::Parser("monsterMap", map);
 
-			x = GameUtils::JsonParser::Parser("x", monsterMapJson).get_int64();
-			y = GameUtils::JsonParser::Parser("y", monsterMapJson).get_int64();
-			centerX = GameUtils::JsonParser::Parser("centerX", monsterMapJson).get_int64();
-			centerY = GameUtils::JsonParser::Parser("centerY", monsterMapJson).get_int64();
+			x = static_cast<int32>(GameUtils::JsonParser::Parser("x", monsterMapJson).get_int64());
+			y = static_cast<int32>(GameUtils::JsonParser::Parser("y", monsterMapJson).get_int64());
+			centerX = static_cast<int32>(GameUtils::JsonParser::Parser("centerX", monsterMapJson).get_int64());
+			centerY = static_cast<int32>(GameUtils::JsonParser::Parser("centerY", monsterMapJson).get_int64());
 
 			room->GetGameMap()->GetMonsterMapInfo()->GetRect().X = x;
 			room->GetGameMap()->GetMonsterMapInfo()->GetRect().Y = y;
@@ -91,8 +91,8 @@ void GameInit::SetMap(boost::json::value& mapJson, boost::json::value& mapUnitJs
 			for(int32 j = 0; j < mapMonsterList.get_array().size(); j++)
 			{
 				auto& unitsInfo = mapMonsterList.get_array()[j];
-				int32 monsterType = GameUtils::JsonParser::Parser("code", unitsInfo).get_int64();
-				int32 monsterCount = GameUtils::JsonParser::Parser("count", unitsInfo).get_int64();
+				int32 monsterType = static_cast<int32>(GameUtils::JsonParser::Parser("code", unitsInfo).get_int64());
+				int32 monsterCount = static_cast<int32>(GameUtils::JsonParser::Parser("count", unitsInfo).get_int64());
 				room->CreateMonster(monsterType, monsterCount);
 			}
 		}
@@ -103,10 +103,10 @@ void GameInit::SetPlayer(boost::json::value& unitJson)
 {
 	for (auto& unit : unitJson.as_array())
 	{
-		int32 code = GameUtils::JsonParser::Parser("code", unit).get_int64();
-		float attack = GameUtils::JsonParser::Parser("attack", unit).get_double();
-		float moveSpeed = GameUtils::JsonParser::Parser("moveSpeed", unit).get_double();
-		int32 hp = GameUtils::JsonParser::Parser("hp", unit).get_int64();
+		int32 code = static_cast<int32>(GameUtils::JsonParser::Parser("code", unit).get_int64());
+		float attack = static_cast<float>(GameUtils::JsonParser::Parser("attack", unit).get_double());
+		float moveSpeed = static_cast<float>(GameUtils::JsonParser::Parser("moveSpeed", unit).get_double());
+		int32 hp = static_cast<int32>(GameUtils::JsonParser::Parser("hp", unit).get_int64());
 
 		GPlayer->GetCharater().emplace(code, Charater(code, attack, moveSpeed, hp));
 	}
@@ -116,11 +116,11 @@ void GameInit::SetMonster(boost::json::value& unitJson)
 {
 	for (auto& unit : unitJson.as_array())
 	{
-		int32 code = GameUtils::JsonParser::Parser("code", unit).get_int64();
-		float attack = GameUtils::JsonParser::Parser("attack", unit).get_double();
-		float moveSpeed = GameUtils::JsonParser::Parser("moveSpeed", unit).get_double();
-		int32 hp = GameUtils::JsonParser::Parser("hp", unit).get_int64();
-		int32 exp = GameUtils::JsonParser::Parser("exp", unit).get_int64();
+		int32 code = static_cast<int32>(GameUtils::JsonParser::Parser("code", unit).get_int64());
+		float attack = static_cast<float>(GameUtils::JsonParser::Parser("attack", unit).get_double());
+		float moveSpeed = static_cast<float>(GameUtils::JsonParser::Parser("moveSpeed", unit).get_double());
+		int32 hp = static_cast<int32>(GameUtils::JsonParser::Parser("hp", unit).get_int64());
+		int32 exp = static_cast<int32>(GameUtils::JsonParser::Parser("exp", unit).get_int64());
 
 		GMonster->GetCharater().emplace(code, Charater(code, attack, moveSpeed, hp, exp));
 	}
@@ -130,8 +130,8 @@ void GameInit::SetWeapon(boost::json::value& unitJson)
 {
 	for (auto& unit : unitJson.as_array())
 	{
-		int32 code = GameUtils::JsonParser::Parser("code", unit).get_int64();
-		float attackSpeed = GameUtils::JsonParser::Parser("attackSpeed", unit).get_double();
+		int32 code = static_cast<int32>(GameUtils::JsonParser::Parser("code", unit).get_int64());
+		float attackSpeed = static_cast<float>(GameUtils::JsonParser::Parser("attackSpeed", unit).get_double());
 
 		GWeapon->GetWeapon().emplace(code, Weapon(code, attackSpeed));
 	}
@@ -141,8 +141,8 @@ void GameInit::SetLv(boost::json::value& lvJson)
 {
 	for (auto expLv : lvJson.get_array())
 	{
-		int32 lv = GameUtils::JsonParser::Parser("lv", expLv).get_int64(); // 레벨
-		int32 exp = GameUtils::JsonParser::Parser("exp", expLv).get_int64(); // 레벨업에 필요한 경험치량
+		int32 lv = static_cast<int32>(GameUtils::JsonParser::Parser("lv", expLv).get_int64()); // 레벨
+		int32 exp = static_cast<int32>(GameUtils::JsonParser::Parser("exp", expLv).get_int64()); // 레벨업에 필요한 경험치량
 
 		GExpLv->GetExpLv().emplace(lv, ExpLv(lv, exp));
 	}
@@ -152,12 +152,12 @@ void GameInit::SetSkill(boost::json::value& unitJson)
 {
 	for (auto& unit : unitJson.as_array())
 	{
-		int32 type = GameUtils::JsonParser::Parser("type", unit).get_int64();
-		float value = GameUtils::JsonParser::Parser("value", unit).get_double();
-		float duration = GameUtils::JsonParser::Parser("duration", unit).get_double();
-		int32 code = GameUtils::JsonParser::Parser("code", unit).get_int64();
-		float coolTime = GameUtils::JsonParser::Parser("coolTime", unit).get_double();
-		int32 weaponCode = GameUtils::JsonParser::Parser("weaponCode", unit).get_int64();
+		int32 type = static_cast<int32>(GameUtils::JsonParser::Parser("type", unit).get_int64());
+		float value = static_cast<float>(GameUtils::JsonParser::Parser("value", unit).get_double());
+		float duration = static_cast<float>(GameUtils::JsonParser::Parser("duration", unit).get_double());
+		int32 code = static_cast<int32>(GameUtils::JsonParser::Parser("code", unit).get_int64());
+		float coolTime = static_cast<float>(GameUtils::JsonParser::Parser("coolTime", unit).get_double());
+		int32 weaponCode = static_cast<int32>(GameUtils::JsonParser::Parser("weaponCode", unit).get_int64());
 
 		GSkill->GetSkill().emplace(code, Skill(type, value, duration, code, coolTime, weaponCode));
 	}
@@ -167,11 +167,11 @@ void GameInit::SetItemEquip(boost::json::value& itemJson)
 {
 	for (auto& equip : itemJson.as_array())
 	{
-		int32 code = GameUtils::JsonParser::Parser("code", equip).get_int64();
-		int32 type = GameUtils::JsonParser::Parser("type", equip).get_int64();
-		int32 gold = GameUtils::JsonParser::Parser("gold", equip).get_int64();
-		int32 attack = GameUtils::JsonParser::Parser("attack", equip).get_int64();
-		int32 speed = GameUtils::JsonParser::Parser("speed", equip).get_int64();
+		int32 code = static_cast<int32>(GameUtils::JsonParser::Parser("code", equip).get_int64());
+		int32 type = static_cast<int32>(GameUtils::JsonParser::Parser("type", equip).get_int64());
+		int32 gold = static_cast<int32>(GameUtils::JsonParser::Parser("gold", equip).get_int64());
+		int32 attack = static_cast<int32>(GameUtils::JsonParser::Parser("attack", equip).get_int64());
+		int32 speed = static_cast<int32>(GameUtils::JsonParser::Parser("speed", equip).get_int64());
 
 		GEquipItem->AddItem(code, type, gold, attack, speed);
 	}
@@ -181,10 +181,10 @@ void GameInit::SetItemEtc(boost::json::value& itemJson)
 {
 	for (auto& etc : itemJson.as_array())
 	{
-		int32 code = GameUtils::JsonParser::Parser("code", etc).get_int64();
-		int32 type = GameUtils::JsonParser::Parser("type", etc).get_int64();
-		int32 gold = GameUtils::JsonParser::Parser("gold", etc).get_int64();
-		int32 maxCnt = GameUtils::JsonParser::Parser("maxCnt", etc).get_int64();
+		int32 code = static_cast<int32>(GameUtils::JsonParser::Parser("code", etc).get_int64());
+		int32 type = static_cast<int32>(GameUtils::JsonParser::Parser("type", etc).get_int64());
+		int32 gold = static_cast<int32>(GameUtils::JsonParser::Parser("gold", etc).get_int64());
+		int32 maxCnt = static_cast<int32>(GameUtils::JsonParser::Parser("maxCnt", etc).get_int64());
 
 		GEtcItem->AddItem(code, type, gold, maxCnt);
 	}
@@ -194,19 +194,19 @@ void GameInit::SetDropItem(boost::json::value& dropJson)
 {
 	for (auto& drop : dropJson.as_array())
 	{
-		int32 monsterCode = GameUtils::JsonParser::Parser("monsterCode", drop).get_int64();
+		int32 monsterCode = static_cast<int32>(GameUtils::JsonParser::Parser("monsterCode", drop).get_int64());
 		boost::json::value dropEquipList = GameUtils::JsonParser::Parser("dropEquipList", drop);
 		for (auto& item : dropEquipList.as_array())
 		{
-			int32 itemCode = GameUtils::JsonParser::Parser("itemCode", item).get_int64();
+			int32 itemCode = static_cast<int32>(GameUtils::JsonParser::Parser("itemCode", item).get_int64());
 			GDropItem->AddDropEquipItem(monsterCode, itemCode);
 		}
-		int32 gold = GameUtils::JsonParser::Parser("gold", drop).get_int64();
+		int32 gold = static_cast<int32>(GameUtils::JsonParser::Parser("gold", drop).get_int64());
 		boost::json::value dropList = GameUtils::JsonParser::Parser("dropList", drop);
 		for (auto& item : dropList.as_array())
 		{
-			int32 itemCode = GameUtils::JsonParser::Parser("itemCode", item).get_int64();
-			int32 cnt = GameUtils::JsonParser::Parser("cnt", item).get_int64();
+			int32 itemCode = static_cast<int32>(GameUtils::JsonParser::Parser("itemCode", item).get_int64());
+			int32 cnt = static_cast<int32>(GameUtils::JsonParser::Parser("cnt", item).get_int64());
 			GDropItem->AddDropItem(monsterCode, itemCode, cnt);
 		}
 		GDropItem->AddGold(monsterCode, gold);
